refactor: move http and typeorm on module products to infra layer
